move coordinator into a separate file
